refactor(rewards page): move connect button to reward table
